Update S3.py with bug fix
In bucket_list_streaming method, it is possible to get empty contents or common_prefixes when the truncation is set to true. In this case a next_marker is returned and I came across such scenario while I was working on a different cloud server which complies with AWS standards. Technically this is within the spec defined by AWS.
“The response might contain fewer keys, but it will never contain more. If there are additional keys that satisfy the search criteria, but these keys were not returned because max-keys was exceeded, the response contains <IsTruncated>true</IsTruncated>. To return the additional keys, see marker.”
So I have added the bug fix to check before accessing the empty lists and to verify if there is next_marker and use it in the next iteration of fetching the response that matches the criteria.